refactor: move error-codes
