+++ .gitignore
@@ -20,3 +20,4 @@ examples/run.js
 /enigma.js*
 /enigma.min*
 /sense-utilities.*
+/error-codes.*
